content/html/content/test/forms/test_input_file_b2g_disabled.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=801635
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 801635</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 801635 **/
  var oldPrefValue = null;
  try {
    oldPrefValue = SpecialPowers.getBoolPref('dom.disable_input_file');
  } catch(e) {
  }

  var input = document.createElement('input');

  SpecialPowers.setBoolPref('dom.disable_input_file', true); 
  input.type='file';
  is(input.type, 'text', "<input type='file'> should be disabled");
  input.type ='text';

  SpecialPowers.setBoolPref('dom.disable_input_file', false); 
  input.type='file';
  is(input.type, 'file', "<input type='file'> should not be disabled");

  SpecialPowers.setBoolPref('dom.disable_input_file', oldPrefValue); 

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=801635">Disable &lt;input type='file'&gt; on B2G</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
